Separate user actions from account actions
